# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  pkg_api_handlers.DeviceAuthorizeRequest:
    description: Device authorization request
    properties:
      user_code:
        example: ABCD-1234
        type: string
    type: object
  pkg_api_handlers.DeviceCodePendingResponse:
    description: Device code pending response
    properties:
      error:
        example: authorization_pending
        type: string
      error_description:
        example: The user has not yet authorized the device
        type: string
    type: object
  pkg_api_handlers.DeviceCodePollRequest:
    description: Device code poll request
    properties:
      device_code:
        example: abc123xyz789...
        type: string
    type: object
  pkg_api_handlers.DeviceCodeResponse:
    description: Device code flow response
    properties:
      device_code:
        example: abc123xyz789...
        type: string
      expires_in:
        example: 900
        type: integer
      interval:
        example: 5
        type: integer
      user_code:
        example: ABCD-1234
        type: string
      verification_uri:
        example: http://localhost:9000/device
        type: string
    type: object
  pkg_api_handlers.ErrorResponse:
    description: Error response returned by the API
    properties:
      details: {}
      error:
        example: bad_request
        type: string
      message:
        example: Invalid request body
        type: string
    type: object
...
  pkg_api_handlers.LogoutRequest:
    description: Logout request
    properties:
      revoke_all:
        example: false
        type: boolean
    type: object
...
  pkg_api_handlers.OAuthCallbackRequest:
    description: OAuth callback request
    properties:
      code:
        example: abc123
        type: string
      state:
        example: xyz789
        type: string
    type: object
  pkg_api_handlers.OAuthStartResponse:
    description: OAuth flow start response
    properties:
      auth_url:
        example: https://github.com/login/oauth/authorize?client_id=...
        type: string
      state:
        example: abc123xyz
        type: string
    type: object
...
  pkg_api_handlers.RefreshTokenRequest:
    description: Refresh token request
    properties:
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
    type: object
...
  pkg_api_handlers.SuccessResponse:
    description: Generic success response
    properties:
      message:
        example: Operation completed successfully
        type: string
    type: object
...
  pkg_api_handlers.TokenResponse:
    description: Authentication token response
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
      expires_at:
        example: "2024-01-15T10:45:00Z"
        type: string
      expires_in:
        example: 900
        type: integer
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
      token_type:
        example: Bearer
        type: string
      user:
        $ref: '#/definitions/pkg_api_handlers.UserResponse'
    type: object
...
  pkg_api_handlers.UserResponse:
    description: User information
    properties:
      avatar_url:
        example: https://avatars.githubusercontent.com/u/1234567
        type: string
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      email:
        example: user@example.com
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        example: John Doe
        type: string
      provider:
        example: github
        type: string
      updated_at:
        example: "2024-01-15T10:30:00Z"
        type: string
    type: object
...
host: localhost:9000
info:
  contact:
    email: support@dagryn.dev
    name: Dagryn Support
    url: https://github.com/mujhtech/dagryn
  description: Local-first, self-hosted developer workflow orchestrator API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Dagryn API
  version: "1.0"
paths:
...
  /api/v1/auth/{provider}:
    get:
      description: Initiates OAuth authentication with the specified provider
      parameters:
      - description: OAuth provider (github, google)
        in: path
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_api_handlers.OAuthStartResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg_api_handlers.ErrorResponse'
      summary: Start OAuth flow
      tags:
      - auth
  /api/v1/auth/{provider}/callback:
    post:
      consumes:
      - application/json
      description: Handles OAuth callback from providers and returns tokens
      parameters:
      - description: OAuth provider (github, google)
        in: path
        name: provider
        required: true
        type: string
      - description: OAuth callback data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/pkg_api_handlers.OAuthCallbackRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_api_handlers.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg_api_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg_api_handlers.ErrorResponse'
      summary: OAuth callback
      tags:
      - auth
  /api/v1/auth/device:
    post:
      description: Initiates device code flow for CLI authentication
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_api_handlers.DeviceCodeResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg_api_handlers.ErrorResponse'
      summary: Request device code
      tags:
      - auth
  /api/v1/auth/device/authorize:
    post:
      consumes:
      - application/json
      description: Authorizes a device code (called by user in browser)
      parameters:
      - description: Device authorize request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/pkg_api_handlers.DeviceAuthorizeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_api_handlers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg_api_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg_api_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Authorize device
      tags:
      - auth
  /api/v1/auth/device/deny:
    post:
      consumes:
      - application/json
      description: Denies a device code authorization
      parameters:
      - description: Device deny request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/pkg_api_handlers.DeviceAuthorizeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_api_handlers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg_api_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg_api_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Deny device
      tags:
      - auth
  /api/v1/auth/device/poll:
    post:
      consumes:
      - application/json
      description: Polls for device code authorization status
      parameters:
      - description: Device code poll request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/pkg_api_handlers.DeviceCodePollRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_api_handlers.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg_api_handlers.ErrorResponse'
        "428":
          description: Authorization pending
          schema:
            $ref: '#/definitions/pkg_api_handlers.DeviceCodePendingResponse'
      summary: Poll device code
      tags:
      - auth
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      description: Revokes the current access token and optionally all tokens
      parameters:
      - description: Logout options
        in: body
        name: body
        schema:
          $ref: '#/definitions/pkg_api_handlers.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_api_handlers.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg_api_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - auth
...
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Exchanges a refresh token for a new access token
      parameters:
      - description: Refresh token request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/pkg_api_handlers.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_api_handlers.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg_api_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg_api_handlers.ErrorResponse'
      summary: Refresh access token
      tags:
      - auth
...
securityDefinitions:
  APIKeyAuth:
    description: API key authentication
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: JWT Bearer token authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"